requirements.txt
Flask
google-cloud-storage
google-api-python-client